% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegMcalt-package.R
\docType{package}
\name{ggsegMcalt-package}
\alias{ggsegMcalt}
\alias{ggsegMcalt-package}
\title{ggsegMcalt: MCALT Atlas for the 'ggseg' Ecosystem}
\description{
Mayo Clinic Adult Lifespan Template (MCALT) brain atlas for the 'ggseg' ecosystem. Provides 'ggseg_atlas' objects for population-specific brain parcellation optimized for aging and Alzheimer's Disease research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegMcalt}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegMcalt/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
