% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mcalt_cortical}
\alias{mcalt_cortical}
\title{MCALT Cortical Atlas}
\usage{
mcalt_cortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical parcellation from the Mayo Clinic Adult Lifespan Template (MCALT)
ADIR122 atlas.
}
\examples{
mcalt_cortical()
\dontrun{plot(mcalt_cortical())}
}
\references{
Schwarz CG, et al. (2017). Alzheimer's & Dementia, 13(7):P1071.
\doi{10.1016/j.jalz.2017.06.1071}
}
\seealso{
Other ggseg_atlases: 
\code{\link{mcalt_subcortical}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
