% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mcalt_subcortical}
\alias{mcalt_subcortical}
\title{MCALT Subcortical Atlas}
\usage{
mcalt_subcortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Subcortical parcellation from the Mayo Clinic Adult Lifespan Template (MCALT)
ADIR122 atlas.
}
\examples{
mcalt_subcortical()
plot(mcalt_subcortical())
}
\references{
Schwarz CG, et al. (2017). Alzheimer's & Dementia, 13(7):P1071.
\doi{10.1016/j.jalz.2017.06.1071}
}
\seealso{
Other ggseg_atlases: 
\code{\link{mcalt_cerebellar}()},
\code{\link{mcalt_cortical}()}
}
\concept{ggseg_atlases}
\concept{subcortical_atlases}
